+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+
+### Added
+
+- Add Verilog: Doctor command to diagnose external tool configuration.
+
 ## [1.23.0] - 2026-05-18
 
 ### Fixed